moving location subpackage into ir subpackage. (#105)
