+++ tests/_test_batch.py
@@ -1,4 +1,4 @@
-import bloqade.location as location
+import bloqade.ir.location as location
 from bloqade.ir import Linear, Constant
 from bloqade.task import HardwareBatch
 import numpy as np
